pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pip>=25.2", # otherwise PyCharm does not detect installed packages
    "pytest-asyncio>=1.1.0",
    "ruff>=0.13.0",
]

load-testing = [
    "locust>=2.40.2",
]

telemetry = [
    "hyperdx-opentelemetry>=0.3.0",
    "opentelemetry-instrumentation-asgi==0.51b0",
    "opentelemetry-instrumentation-asyncio==0.51b0",
    "opentelemetry-instrumentation-click==0.51b0",
    "opentelemetry-instrumentation-dbapi==0.51b0",
    "opentelemetry-instrumentation-fastapi==0.51b0",
    "opentelemetry-instrumentation-grpc==0.51b0",
    "opentelemetry-instrumentation-httpx==0.51b0",
    "opentelemetry-instrumentation-logging==0.51b0",
    "opentelemetry-instrumentation-psycopg==0.51b0",
    "opentelemetry-instrumentation-requests==0.51b0",
    "opentelemetry-instrumentation-sqlalchemy==0.51b0",
    "opentelemetry-instrumentation-sqlite3==0.51b0",
    "opentelemetry-instrumentation-starlette==0.51b0",
    "opentelemetry-instrumentation-system-metrics==0.51b0",
    "opentelemetry-instrumentation-threading==0.51b0",
    "opentelemetry-instrumentation-tortoiseorm==0.51b0",
    "opentelemetry-instrumentation-urllib==0.51b0",
    "opentelemetry-instrumentation-urllib3==0.51b0",
    "opentelemetry-instrumentation-wsgi==0.51b0",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'aarch64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
]
